requirements.txt
# Django and core dependencies
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
drf-yasg==1.21.7

# Authentication and Security
rest-framework-simplejwt==5.3.0
pyotp==2.9.0
qrcode==7.4.2
cryptography==42.0.5

# Database
psycopg2-binary==2.9.9

# Celery for background tasks
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Image processing
Pillow==10.2.0

# Email
django-anymail==10.2

# Monitoring and logging
sentry-sdk==1.40.6

# Development tools
django-debug-toolbar==4.2.0
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
django-redis==5.4.0

# Utilities
python-decouple==3.8
django-environ==0.11.2
